+++ src/main.ts
@@ -1,6 +1,6 @@
-import CalendarSidebarPlugin from './plugin';
+import DaylinePlugin from './plugin';
 
 declare const module: { exports: unknown };
 
 // Obsidian loads main.js as a CommonJS plugin class.
-module.exports = CalendarSidebarPlugin;
+module.exports = DaylinePlugin;